Pull request #3830: [BLUETOOTH-16957] Fix MISRA issue
Merge in CONNINT/bluetooth_public from bugfix/BLUETOOTH-16957-kw47-kw45-26.03.00-hcit_generic_adapter_interface_misra to develop/main

Squashed commit of the following:

commit f5ebc18a046b71f70dcf8a92d2025972824612b1
Author: Andrei Istodorescu <andrei.istodorescu@nxp.com>
Date:   Mon Feb 23 11:52:05 2026 +0200

    [BLUETOOTH-16957] Fix MISRA issue

+++ hci_transport/source/hcit_generic_adapter_interface.c
@@ -279,9 +279,9 @@ bleResult_t Hcit_SendPacket
             uint32_t payloadLen = 0U;
             aclDataPkt = true;
             payloadLen = (((((uint32_t)buf[3U]) << 8U) | ((uint32_t)buf[2U])));
-            if (payloadLen <= (UINT32_MAX - 5U))
+            if (payloadLen <= ((uint32_t)UINT16_MAX - 5U))
             {
-              totalLen = payloadLen + 5U;
+                totalLen = payloadLen + 5U;
             }
         }
         else
